Refactor: Comment out drawing functions in painters
- Commented out `_drawBMIRanges` in `bmi_chart_painter.dart`.
- Corrected type in `chart_data_point_drawer.dart`.
- Commented out `_rangeDrawer.drawReferenceRanges` in `blood_pressure_chart_painter.dart`.

+++ lib/blood_pressure/Drawer/chart_data_point_drawer.dart
@@ -481,7 +481,7 @@ class ChartDataPointDrawer {
 
     // Ease-out cubic for smoother finish
     final t = ((animation.value - delay) / duration).clamp(0.0, 1.0);
-    return 1.0 - pow(1.0 - t, 3) as double;
+    return 1.0 - pow(1.0 - t, 3);
   }
 
   // Helper method for position calculation
